refactor(parsing): разделён модуль парсинга RSS и Atom
FeedParserService оставлен facade для выбора parser flow и публичных entry points
XML document builder, feed kind detection, RSS parser, Atom parser, parser models и parse pipeline diagnostics вынесены в отдельные файлы
поведение parsing pipeline сохранено, но границы XML parsing, DTO extraction и diagnostics стали явнее
