+++ README.md
@@ -529,7 +529,7 @@
 - [x] `Source Management State Partition`: разрезать `SourceManagementScreenState` на отдельные state/reducer файлы для `Add Feed`, `Move Source` и `Create Folder`, оставив root state только координатором presentation destination и derived view state; сохранить текущий behavior, но сделать каждый flow тестируемым и review-friendly отдельно;
 - [x] `Source Management View Split`: разделить крупный `SourceManagementScreenView` на root navigation container, destination factory и отдельные view-файлы для add/edit feed, move source и create/edit folder, чтобы UI-слой не смешивал navigation plumbing, bindings и конкретную разметку форм;
 - [x] `Reader Content Rendering Pipeline Split`: разнести `ArticleScreenContentRenderer` на доменные части: body block models, HTML block tokenizer / parser, inline text renderer, media URL resolver и final body assembly; renderer должен остаться thin orchestration layer между `ArticleScreenBodyPayloadNormalizer` и UI presentation blocks;
-- [ ] `Feed Parsing Module Split`: разделить `FeedParserService` на XML document builder, feed kind detection, RSS parser, Atom parser и parse pipeline diagnostics, чтобы будущие tests для parsing / normalizer не зависели от одного большого файла с несколькими уровнями ответственности;
+- [x] `Feed Parsing Module Split`: разделить `FeedParserService` на XML document builder, feed kind detection, RSS parser, Atom parser и parse pipeline diagnostics, чтобы будущие tests для parsing / normalizer не зависели от одного большого файла с несколькими уровнями ответственности;
 - [ ] `Refresh Pipeline Boundary Cleanup`: проверить `FeedRefreshService` после накопленных archive / retention / diagnostics изменений и выделить из него отдельные шаги conditional request, not-modified handling, fetched response reconciliation, diagnostics/log persistence и batch orchestration, не меняя публичный `FeedRefreshCoordinating` contract без необходимости;
 - [ ] `Loading Indicator Contract Audit`: пройти оставшиеся loading states после `Shared Loading Indicator Tint And Reuse` и явно зафиксировать, где допустим системный `ProgressView` / UIKit indicator, а где должен использоваться `AppRefreshIndicator`, чтобы UI не возвращался к смешанным indicator semantics перед UI tests;
 - [ ] `Test Harness Decomposition`: разложить `RSSReaderTestSupport` на focused builders / fixtures: SwiftData harness, scripted HTTP client, DTO factories, logging doubles и source-management fixtures; тесты должны импортировать общие helpers без неявной зависимости от feed refresh harness;
